requirements.txt
Flask
gunicorn
sqlalchemy
mysql-connector